requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic[email]==2.5.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
pandas==2.1.3
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
aiosmtplib==3.0.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
